+++ tests/memory_tests.cpp
@@ -17,6 +17,8 @@
 #include "test_fixtures.h"
 #include "rmm/rmm.h"
 
+#include <numeric>
+
 // Helper macros to simplify testing for success or failure
 #define ASSERT_SUCCESS(res) ASSERT_EQ(RMM_SUCCESS, (res));
 #define ASSERT_FAILURE(res) ASSERT_NE(RMM_SUCCESS, (res));
@@ -44,10 +46,11 @@ struct MemoryManagerTest :
     static void SetUpTestCase() {
         ASSERT_FALSE(rmmIsInitialized(0));
         ASSERT_EQ( cudaSuccess, cudaStreamCreate(&stream) );
-        rmmOptions_t options {allocationMode(), 0, false};
-        ASSERT_SUCCESS( rmmInitialize(&options) );
-        rmmOptions_t options_set;
+        rmmOptions_t options {};
+        options.allocation_mode = allocationMode();
+        ASSERT_SUCCESS(rmmInitialize(&options));
         // verify initialized
+        rmmOptions_t options_set;
         ASSERT_TRUE(rmmIsInitialized(&options_set));
         // verify initialized options
         ASSERT_EQ(options_set.allocation_mode, options.allocation_mode);
@@ -61,6 +64,7 @@ struct MemoryManagerTest :
         ASSERT_EQ( cudaSuccess, cudaStreamDestroy(stream) );
     }
 };
+
 
 template <rmmAllocationMode_t mode>
 struct ModeType {
@@ -229,6 +233,141 @@ TYPED_TEST(MemoryManagerTest, GetInfo) {
 }
 
 TYPED_TEST(MemoryManagerTest, AllocationOffset) {
+    char *a = nullptr, *b = nullptr;
+    ptrdiff_t offset = -1;
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
+    ASSERT_SUCCESS( RMM_ALLOC(&b, size_mb, stream) );
+
+    ASSERT_SUCCESS( rmmGetAllocationOffset(&offset, a, stream) );
+    ASSERT_GE(offset, 0);
+
+    ASSERT_SUCCESS( rmmGetAllocationOffset(&offset, b, stream) );
+    ASSERT_GE(offset, 0);
+
+    ASSERT_SUCCESS( RMM_FREE(a, stream) );
+    ASSERT_SUCCESS( RMM_FREE(b, stream) );
+}
+
+template <typename T>
+struct MultiGPUMemoryManagerTest : public MemoryManagerTest<T> {
+    static void SetUpTestCase() {
+        ASSERT_FALSE(rmmIsInitialized(0));
+        ASSERT_EQ( cudaSuccess, cudaStreamCreate(&stream) );
+        int device_count{};
+        ASSERT_EQ(cudaSuccess, cudaGetDeviceCount(&device_count));
+        std::vector<int> devices(device_count);
+        std::iota(begin(devices), end(devices),0);
+        rmmOptions_t options{};
+        options.allocation_mode = T::alloc_mode;
+        options.devices = devices;
+        ASSERT_SUCCESS( rmmInitialize(&options) );
+        rmmOptions_t options_set;
+        // verify initialized
+        ASSERT_TRUE(rmmIsInitialized(&options_set));
+        // verify initialized options
+        ASSERT_EQ(options_set.allocation_mode, options.allocation_mode);
+        ASSERT_EQ(options_set.initial_pool_size, options.initial_pool_size);
+        ASSERT_EQ(options_set.enable_logging, options.enable_logging);
+    }
+};
+
+TYPED_TEST_CASE(MultiGPUMemoryManagerTest, allocation_modes);
+
+// Init / Finalize tests
+
+TYPED_TEST(MultiGPUMemoryManagerTest, Initialize) {
+    
+    // Initialized in Fixture class.
+    rmmOptions_t options;
+    ASSERT_TRUE(rmmIsInitialized(&options));
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, Finalize) {
+    // Empty because handled in Fixture class.
+}
+
+// zero size tests
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocateZeroBytes) {
+    char *a = 0;
+    ASSERT_SUCCESS( RMM_ALLOC(&a, 0, stream) );
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, NullPtrAllocateZeroBytes) {
+    char ** p{nullptr};
+    ASSERT_SUCCESS( RMM_ALLOC(p, 0, stream) );
+}
+
+// Bad argument tests
+
+TYPED_TEST(MultiGPUMemoryManagerTest, NullPtrInvalidArgument) {
+    char ** p{nullptr};
+    rmmError_t res = RMM_ALLOC(p, 4, stream);
+    ASSERT_FAILURE(res);
+    ASSERT_EQ(RMM_ERROR_INVALID_ARGUMENT, res);
+}
+
+// Simple allocation / free tests
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocateWord) {
+    char *a = 0;
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_word, stream) );
+    ASSERT_SUCCESS( RMM_FREE(a, stream) );
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocateKB) {
+    char *a = 0;
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_kb, stream) );
+    ASSERT_SUCCESS( RMM_FREE(a, stream) );
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocateMB) {
+    char *a = 0;
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );
+    ASSERT_SUCCESS( RMM_FREE(a, stream) );
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocateGB) {
+    char *a = 0;
+    ASSERT_SUCCESS( RMM_ALLOC(&a, size_gb, stream) );
+    ASSERT_SUCCESS( RMM_FREE(a, stream) );
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocateTB) {
+    char *a = 0;
+    size_t freeBefore = 0, totalBefore = 0;
+    ASSERT_SUCCESS( rmmGetInfo(&freeBefore, &totalBefore, stream) );
+
+    if ((this->allocationMode() & CudaManagedMemory) || 
+        (size_tb < freeBefore)) {
+        // TODO investigate and fix this
+        //ASSERT_SUCCESS( RMM_ALLOC(&a, size_tb, stream) );
+    }
+    else {
+        ASSERT_FAILURE( RMM_ALLOC(&a, size_tb, stream) );
+        ASSERT_FAILURE( RMM_FREE(a, stream) );
+    }
+    
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocateTooMuch) {
+    char *a = 0;
+    ASSERT_FAILURE( RMM_ALLOC(&a, size_pb, stream) );
+    ASSERT_FAILURE( RMM_FREE(a, stream) );
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, FreeZero) {
+    ASSERT_SUCCESS( RMM_FREE(0, stream) );
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, GetInfo) {
+    size_t freeBefore = 0, totalBefore = 0;
+    ASSERT_SUCCESS( rmmGetInfo(&freeBefore, &totalBefore, stream) );
+    ASSERT_GE(freeBefore, 0);
+    ASSERT_GE(totalBefore, 0);
+}
+
+TYPED_TEST(MultiGPUMemoryManagerTest, AllocationOffset) {
     char *a = nullptr, *b = nullptr;
     ptrdiff_t offset = -1;
     ASSERT_SUCCESS( RMM_ALLOC(&a, size_mb, stream) );